chore(quickcheck): Refactor quicktests in tests folder
