+++ crates/sigmf/Makefile
@@ -1,0 +1,9 @@
+test: cargo-test quickcheck
+
+cargo-test:
+	cargo test
+
+quickcheck:
+	./quickcheck-test.sh
+
+.PHONY: quickcheck cargo-test
